M1/requirements.txt
fastapi==0.101.1
numpy==1.24.4
opencv-python==4.8.0
Pillow==9.0.0
rembg==2.0.50
supervision==0.12.0
torch==1.12.1+cpu
pyngrok==6.0.0
segment_anything==1.0
pathlib==1.0.1
